Add a region utility class

+++ src/io/github/zachohara/percussionpacker/cardspace/ColumnSeparator.java
@@ -28,7 +28,7 @@ public class ColumnSeparator extends Pane implements MouseHandler {
 	
 	public static final String STYLE = "-fx-background-color: black;"
 			+ "-fx-background-radius: 3 3 3 3";
-	public static final int THICKNESS = 3; // in pixels
+	public static final double THICKNESS = 3; // in pixels
 	
 	private ColumnPane parent;
